Add files via upload
Base excel sheet containing the motors  and gearboxes. The gearboxes are separated into three sections: first gearbox, second gearbox, and third gearbox. The combinations do not need to use all three but it is REQUIRED to use the third gearbox. This is because the third gearbox is a right angle gearbox that is needed based on the way the sponsors build the door automation systems.